fix warnings

+++ Editor/Scripts/Preview/PreviewClipInfo.cs
@@ -7,7 +7,7 @@ internal struct PreviewClipInfo {
     public double Duration;
     public double TimeScale;
     public double ClipIn;
-    public float FramePerSecond;
+    public double FramePerSecond;
     public float ImageDimensionRatio;
 
     //What is visible on the TimelineWindow
